+++ src/test/java/br/com/academico/endereco/EnderecoServiceTest.java
@@ -25,41 +25,41 @@ public class EnderecoServiceTest {
 
     @Test
     public void teste_recuperar_endereco_por_id(){
-        Endereco endereco = enderecoService.recuperar(10);
+        Endereco endereco = enderecoService.recuperar(10L);
         assertTrue("O retorno do método recuperar deve ser um objeto Endereco: ", endereco instanceof Endereco);
     }    
 
     @Test
     public void teste_criar_endereco() {
-        Endereco endereco = new Endereco(33333, "Rua H", "Centro", "Tobias Barreto", "Sergipe");
-        int idEndereco = enderecoService.criar(endereco);
-        assertTrue("O retorno do método criar deve ser um ID de um Endereco criado: ", idEndereco == (int)idEndereco);
+        Endereco endereco = new Endereco(33333L, "Rua H", "Centro", "Tobias Barreto", "Sergipe");
+        Long idEndereco = enderecoService.criar(endereco);
+        assertTrue("O retorno do método criar deve ser um ID de um Endereco criado: ", idEndereco == (Long)idEndereco);
     }
 
     @Test
     public void teste_atualizar_endereco_por_id(){
-        Endereco endereco = new Endereco(44444, "Rua N", "Centro", "Tobias Barreto", "Sergipe");
-        Endereco enderecoAtualizado = enderecoService.atualizar(12, endereco);
+        Endereco endereco = new Endereco(44444L, "Rua N", "Centro", "Tobias Barreto", "Sergipe");
+        Endereco enderecoAtualizado = enderecoService.atualizar(12L, endereco);
         assertTrue("O retorno do método atualizar deve ser um objeto Endereco: ", enderecoAtualizado instanceof Endereco);
     }
 
     @Test
     public void teste_deletar_endereco_por_id(){
-        int idEnderecoDeletado = enderecoService.deletar(587);
-        assertTrue("O retorno do método deletar deve ser um ID do Endereco deletado: ", idEnderecoDeletado == (int)idEnderecoDeletado);
+        Long idEnderecoDeletado = enderecoService.deletar(587L);
+        assertTrue("O retorno do método deletar deve ser um ID do Endereco deletado: ", idEnderecoDeletado == (Long)idEnderecoDeletado);
     }
 
     
     @Test
     public void teste_atlerar_status_endereco_por_id(){
-        Endereco enderecoAtualizado = enderecoService.mudarStatus(12, StatusEndereco.DESATIVO);
+        Endereco enderecoAtualizado = enderecoService.mudarStatus(12L, StatusEndereco.DESATIVO);
         assertTrue("O retorno do método mudar status deve ser um objeto Endereco: ", enderecoAtualizado instanceof Endereco);
     }
 
     @Test
     public void teste_recuperar_endereco_por_id_inexistente(){
         Exception exception = assertThrows(EnderecoNaoExisteException.class, () -> {
-            Endereco endereco = enderecoService.recuperar(999);
+            Endereco endereco = enderecoService.recuperar(999L);
         });     
         String mensagemEsperada = "O endereço não existe na base de dados.";
         String MensagemLancada = exception.getMessage();
@@ -68,9 +68,9 @@ public class EnderecoServiceTest {
 
     @Test
     public void teste_atualizar_endereco_por_id_inexistente(){
-        Endereco endereco = new Endereco(44444, "Rua N", "Centro", "Tobias Barreto", "Sergipe");
+        Endereco endereco = new Endereco(44444L, "Rua N", "Centro", "Tobias Barreto", "Sergipe");
         Exception exception = assertThrows(EnderecoNaoExisteException.class, () -> {
-            Endereco enderecoAtualizado = enderecoService.atualizar(999, endereco);
+            Endereco enderecoAtualizado = enderecoService.atualizar(999L, endereco);
         });     
         String MensagemEsperada = "O endereço não existe na base de dados.";
         String MensagemLancada = exception.getMessage();
@@ -79,9 +79,9 @@ public class EnderecoServiceTest {
 
     @Test
     public void teste_criar_endereco_cep_invalido(){
-        Endereco endereco = new Endereco(88888, "Rua H", "Centro", "Tobias Barreto", "Sergipe");
+        Endereco endereco = new Endereco(88888L, "Rua H", "Centro", "Tobias Barreto", "Sergipe");
         Exception exception = assertThrows(CEPEnderecoInvalidoException.class, () -> {
-            int idEndereco = enderecoService.criar(endereco);
+            Long idEndereco = enderecoService.criar(endereco);
         });     
         String MensagemEsperada = "O CEP do endereço é inválido.";
         String MensagemLancada = exception.getMessage();
